Responding to feedback:
- Moved constants and exception to events module.
- Created requester module to hold execute_request function.
